admin/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-bootstrap==0.2.4
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-mysql==4.7.1
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.1.1
pamqp==2.3.0
pika==1.3.0
pip==22.2.2
pycodestyle==2.8.0
pytz==2022.2.1
PyYAML==6.0
rabbitpy==2.0.1
requests==2.28.1
setuptools==63.4.1
sqlparse==0.4.1
toml==0.10.2
tzdata==2021.1
urllib3==1.26.12
wheel==0.37.1
wincertstore==0.2
